feat(terms-and-conditions): uses terms-and-conditions component as page
